backend/requirements.txt
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
pyotp<=2.8.0
channels==3.0.5
PyJWT<=2.6.0
google-auth<=2.17.3
requests<=2.29.0
django-filter<=23.2
drf-nested-forms<=1.1.7
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
python-dotenv==1.0.0
sqlparse==0.4.4
uvloop==0.17.0
djangorestframework-simplejwt==5.3.1
daphne<=4.0.0
gunicorn==21.2.0
uwsgi>=2.0.20,<2.1
bleach<=6.1.0